requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0
pyarrow>=10.0.0

# Configuration and file handling
PyYAML>=6.0
python-dotenv>=0.19.0

# Data visualization and dashboard
streamlit>=1.25.0
plotly>=5.15.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0

# Statistical analysis and anomaly detection
scipy>=1.9.0
scikit-learn>=1.1.0

# Memory and performance monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality and formatting
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
isort>=5.10.0

# Development and debugging
jupyter>=1.0.0
ipython>=8.0.0

# File format support
openpyxl>=3.0.0
xlrd>=2.0.0

# Progress bars and utilities
tqdm>=4.64.0
click>=8.0.0

# Date and time handling
python-dateutil>=2.8.0

# Email validation
email-validator>=1.3.0

# Logging enhancements
colorlog>=6.7.0